Revert "rename file direct line tracer"
This reverts commit 506fcb3676c0b173c9a6fb4bd8fc28443c576150.
